requirements.txt
google-adk[extensions]==1.21.0
litellm==1.80.5
PyYAML==6.0.3
azure-identity==1.25.1
azure-storage-blob==12.27.1
beautifulsoup4==4.13.4
fastapi==0.123.10
markdown2==2.5.4
numpy==2.2.6
openai==2.14.0
pre-commit==4.5.1
pydantic==2.12.5
pytest-cov==7.0.0
pytest==9.0.2
python-dotenv==1.2.1
requests==2.32.5
ruff==0.14.10
typing-extensions==4.15.0
unstructured[pdf,docx]==0.18.21
wikipedia==1.4.0